Added PacketSerializer class
